Move shared test helpers out of conftest and document test-helper placement
Importing from a conftest.py gives it a second module identity and can double-register its
fixtures. Rename the package conftest.py to helpers.py (a plain module) and repoint every
import; the package needs no conftest since it defines no fixtures.

Add ddev/tests/AGENTS.md documenting the conftest runtime rule (nothing imports from conftest
except fixture types under TYPE_CHECKING) and the helper scope/placement scheme (local
helpers.py, subtree helpers/, global tests/helpers/; promote never copy; imports point upward).
This lives under ddev/tests rather than the repo-root AGENTS.md because the scope scheme is
ddev-specific.
